+++ module/data/active-effect.js
@@ -20,7 +20,12 @@ export class TorgActiveEffectData extends (foundry.data.ActiveEffectTypeDataMode
     const schema = (game.release.generation >= 14) ? foundry.data.ActiveEffectTypeDataModel.defineSchema() : {};
     Object.assign(schema,
       {
-        // ...foundry.data.ActiveEffectTypeDataModel.defineSchema(),    // Foundry 14+
+        applyOnOutcome: new fields.StringField({
+          choices: CONFIG.torgeternity.testOutcomeLabel,
+          required: true,
+          initial: '',
+          blank: true
+        }),
         transferOnOutcome: new fields.StringField({
           choices: CONFIG.torgeternity.testOutcomeLabel,
           required: true,
@@ -36,11 +41,16 @@ export class TorgActiveEffectData extends (foundry.data.ActiveEffectTypeDataMode
           initial: 'target',
           blank: false
         }),
-        activeIfTrait: newTraitsField(),
-        applyIfAttackTrait: newTraitsField(),
-        applyIfDefendTrait: newTraitsField(),
-        defendAgainstTrait: newTraitsField(),
+        activeIfTrait: newTraitsField('effectActiveTraits'),
+        applyIfAttackTrait: newTraitsField('effectTestTraits'),
+        applyIfDefendTrait: newTraitsField('effectTestTraits'),
+        defendAgainstTrait: newTraitsField('effectTestTraits'),
+        applyIfAttackTraitCombine: newCombineTraitsField(),
+        applyIfDefendTraitCombine: newCombineTraitsField(),
+        defendAgainstTraitCombine: newCombineTraitsField(),
+
         combatToggle: new fields.BooleanField({ initial: false, }),
+        concentratingId: new fields.DocumentUUIDField({ nullable: true })
       })
     return schema;
   }
@@ -70,7 +80,7 @@ export class TorgActiveEffectData extends (foundry.data.ActiveEffectTypeDataMode
    */
   get isSuppressed() {
     // Don't apply the AE to the owning actor if it is being transferred on an attack
-    if (this.transferOnOutcome.length || this.defendAgainstTrait.size) return true;
+    if (this.applyOnOutcome.length || this.transferOnOutcome.length || this.defendAgainstTrait.size) return true;
 
     // If the trait is conditionally active, then check for traits/conditions on the owning actor (if any)
     if (!this.activeIfTrait.size) return false;
@@ -84,4 +94,18 @@ export class TorgActiveEffectData extends (foundry.data.ActiveEffectTypeDataMode
       (!item.system.canEquip || item.isEquipped) &&
       item.system.traits.find(trait => this.activeIfTrait.has(trait)));
   }
+}
+
+function newCombineTraitsField() {
+  return new fields.StringField({
+    blank: false,
+    nullable: false,
+    required: true,
+    choices: {
+      'and': "torgeternity.activeEffect.match.and",
+      'or': "torgeternity.activeEffect.match.or",
+    },
+    trim: true,
+    initial: "or",
+  })
 }