+++ module/settings.js
@@ -17,15 +17,20 @@ export function registerTorgSettings() {
     default: true, // The default value for the setting
   });
 
-  game.settings.register('torgeternity', 'y1pp', {
-    name: 'torgeternity.settingMenu.useYear1PPTent.name',
-    hint: 'torgeternity.settingMenu.useYear1PPTent.hint',
-    scope: 'world',
-    config: true,
-    type: Boolean,
-    default: false,
-    requiresReload: true,
-  });
+  if (CONFIG.torgeternity.axiomByCosm.panPacifica) {
+    game.settings.register('torgeternity', 'y1pp', {
+      name: 'torgeternity.settingMenu.useYear1PPTent.name',
+      hint: 'torgeternity.settingMenu.useYear1PPTent.hint',
+      scope: 'world',
+      config: true,
+      type: Boolean,
+      default: false,
+      requiresReload: true,
+    });
+    if (game.settings.get('torgeternity', 'y1pp')) {
+      CONFIG.torgeternity.axiomByCosm.panPacifica.spirit = 10;
+    }
+  }
 
   // ---------Set up Cards
   game.settings.register('torgeternity', 'setUpCards', {